% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadChirpsRainfall.R
\name{downloadChirpsRainfall}
\alias{downloadChirpsRainfall}
\title{Download Chirps Rainfall daily data by year}
\usage{
downloadChirpsRainfall(dir_out, years, aoi)
}
\arguments{
\item{dir_out}{character. Directory where you want to save the raster images that you are going to download.}

\item{years}{numeric. The period in years that the function should download images.}

\item{aoi}{spatVector object. Provides the boundaries where chirps data should be limited to (datum = WGS84).}
}
\value{
raster files of each year containing daily layers for each (1 for each day of given year).
}
\description{
Download Chirps Rainfall daily data by year based on area of interest
}
\examples{


\dontrun{

require(terra)

area_of_interest = vect("./paracatu.shp")


downloadChirpsRainfall(dir_out = "./temp/chirpsRainfall",
                    years = c(1990:1991),
                   aoi = area_of_interest)

}


}
\references{
https://cran.r-project.org/web/packages/chirps/chirps.pdf
}
