% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTerraClimateParallel.R
\name{downloadTerraClimateParallel}
\alias{downloadTerraClimateParallel}
\title{Download terraClimate monthly data}
\usage{
downloadTerraClimateParallel(
  aoi,
  dir_out,
  variable,
  years,
  ncores = 1,
  retry = 10,
  timeout = 600
)
}
\arguments{
\item{aoi}{spatVector object. Provides the boundaries where chirps data should be limited to (datum = WGS84).}

\item{dir_out}{character. Directory where you want to save the raster images that you are going to download.}

\item{variable}{character. Variable to download. See details for more information.}

\item{years}{numeric. The period in years that the function should download images.}

\item{ncores}{numeric. numeric. The number of processor cores to use for parallelizing the download operation. Default is 1 (no parallelization).}

\item{retry}{numeric. numeric. The number of retry attempts for failed downloads. Default is 100.}

\item{timeout}{numeric.numeric. The timeout in seconds for each download attempt. Default is 600.}
}
\value{
raster files of each year containing 12 layers each (1 for each month of given year).
}
\description{
Download TerraClimate monthly data based on area of interest
}
\details{
\itemize{
\item Variable descriptions:
}

aet (Actual Evapotranspiration, monthly total), units = mm

def (Climate Water Deficit, monthly total), units = mm

pet (Potential evapotranspiration, monthly total), units = mm

ppt (Precipitation, monthly total), units = mm

q (Runoff, monthly total), units = mm

soil (Soil Moisture, total column - at end of month), units = mm

srad (Downward surface shortwave radiation), units = W/m2

swe (Snow water equivalent - at end of month), units = mm

tmax (Max Temperature, average for month), units = C

tmin (Min Temperature, average for month), units = C

vap (Vapor pressure, average for month), units  = kPa

ws (Wind speed, average for month), units = m/s

vpd (Vapor Pressure Deficit, average for month), units = kpa

PDSI (Palmer Drought Severity Index, at end of month), units = unitless
}
\examples{

\dontrun{

require(terra)

area_of_interest = vect("./paracatu.shp")


pptAoi = downloadTerraClimateParallel(dir_out = "./temp/terraClimate",
                     variable = "ppt",
                    years = c(1990:2000),
                    ncores = 5,
                   aoi = area_of_interest)



#set sequential plan
future::plan(future::sequential)

}

}
\references{
adapted from download_terraclimate function of cropDemand package.

https://search.r-project.org/CRAN/refmans/cropDemand/html/00Index.html

Abatzoglou, J.T., S.Z. Dobrowski, S.A. Parks, K.C. Hegewisch, 2018, Terraclimate, a high-resolution global dataset of monthly climate and climatic water balance from 1958-2015, Scientific Data,
}
