% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag1AutoCorrelation.R
\name{lag1AutoCorrelation}
\alias{lag1AutoCorrelation}
\title{Lag 1 AutoCorrelation}
\usage{
lag1AutoCorrelation(seriesStatisticsResultSeries)
}
\arguments{
\item{seriesStatisticsResultSeries}{list of tibbles containing annual or monthly series of all stations;
(series output from \code{\link[=seriesStatistics]{seriesStatistics()}} function).}
}
\value{
tibble with autocorrelation based on lag1 (CorrelationLag1 column), absolute confidente interval and logical column indicating if station series is autocorrelated
}
\description{
Evaluate serial correlation (or autocorrelation) with 1 lag.
}
\examples{

\dontrun{


# Fech a inventory of fluviometric stations for the state of Minas Gerais

inv <- inventory(
  states = "MINAS GERAIS",
  stationType = "flu",
  as_sf = TRUE,
  aoi = NULL)

# Download the first 10 stations from the inventory

s_data <- stationsData(
  inventoryResult = inv[1:10,],
  deleteNAstations = TRUE)

# Organize the data for the stations

org_data <- organize(
  stationsDataResult = s_data
)

# Filter the data for desired period and quality contorl

final_data <- selectStations(
  stationsDataResult = org_data,
  mode = "yearly",
  maxMissing = 10,
  minYears = 15,
  month = 1,
  iniYear = NULL,
  finYear = NULL,
  consistedOnly = FALSE,
  plot = TRUE
)

# Annual mean stream flow serie for each station
Qmean_years = seriesStatistics(final_data, statistics = "Qmean")


#lag1Autocorrelation

autoCorr = lag1AutoCorrelation(Qmean_years$series)

}




}
\references{
Based on acf function
(https://www.rdocumentation.org/packages/forecast/versions/8.21.1/topics/Acf)

Fundamentals of Statistical Hydrology
(https://link.springer.com/book/10.1007/978-3-319-43561-9)

The Mann-Kendall Test Modified by Effective Sample Size to Detect Trend in Serially Correlated Hydrological Series
(https://link.springer.com/article/10.1023/B:WARM.0000043140.61082.60)
}
