% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mMonthlyPlot.R
\encoding{UTF-8}
\name{mMonthlyPlot}
\alias{mMonthlyPlot}
\title{Plot streamflow or rainfall monthly average graph}
\usage{
mMonthlyPlot(
  organizeResult,
  maxMissing = 10,
  minYears = 15,
  consistedOnly = TRUE
)
}
\arguments{
\item{organizeResult}{list, tibble data frame; provides a list containing
the data frames of raw records for each station downloaded from ANA web API
(output from \code{\link[=organize]{organize()}} function).}

\item{maxMissing}{numeric; indicates the maximum threshold of missing data allowed at each year.
The default is 10 percent.}

\item{minYears}{numeric; indicates the minimum years of complete data allowed. The
default is 15 years.}

\item{consistedOnly}{logical; should only consisted data be considered?
The default is TRUE.}
}
\value{
Saved plot.
}
\description{
Plot streamflow or rainfall monthly average to help users identify hydrological year
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Fech a inventory of fluviometric stations for the state of Minas Gerais

inv <- inventory(
  states = "MINAS GERAIS",
  stationType = "flu",
  as_sf = TRUE,
  aoi = NULL
)

# Download the first 10 stations from the inventory

s_data <- stationsData(
  inventoryResult = inv[1:10,],
  deleteNAstations = TRUE
)

# Organize the data for the stations

org_data <- organize(
  stationsDataResult = s_data
)

# Filter the data for desired period and quality control

mMonthlyPlot(
  stationsDataResult = org_data,
  maxMissing = 10,
  minYears = 15,
  consistedOnly = FALSE
)
\dontshow{\}) # examplesIf}
}
