% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mMonthlyStat.R
\name{mMonthlyStat}
\alias{mMonthlyStat}
\title{Mean and Median Monthly Statistics}
\usage{
mMonthlyStat(selectionResultSeries)
}
\arguments{
\item{selectionResultSeries}{list of tibble; lista de tibbles obtidos com a função \link{selectStations} contendo a série de dados diários.}
}
\value{
dataframe contendo o início do mês úmido e seco.
}
\description{
Evaluate Mean and Median monthly statistics based \link{selectStations} function series result
}
\details{
Computa estatística mensal (média ou mediana) nos meses do ano e, posteriormente, e realiza uma média movel de 6 meses para identificar período seco e úmido no ano.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Fech a inventory of fluviometric stations for the state of Minas Gerais

inv <- inventory(
  states = "MINAS GERAIS",
  stationType = "flu",
  as_sf = TRUE,
  aoi = NULL
)

# Download the first 10 stations from the inventory

s_data <- stationsData(
  inventoryResult = inv[1:10,],
  deleteNAstations = TRUE
)

# Organize the data for the stations

org_data <- organize(stationsDataResult = s_data)

# Filter the data for desired period and quality contorl

final_data <- selectStations(
  organizeResult = org_data,
  mode = "yearly",
  maxMissing = 10,
  minYears = 15,
  month = 1,
  iniYear = NULL,
  finYear = NULL,
  consistedOnly = FALSE,
  plot = TRUE
)

mStats = mMonthlyStat(final_data)

\dontshow{\}) # examplesIf}
}
