% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\encoding{UTF-8}
\name{organize}
\alias{organize}
\title{Organizes the raw data provided by ANA for its stations}
\usage{
organize(stationsDataResult)
}
\arguments{
\item{stationsDataResult}{list, tibble data frame; provides a list containing
the data frames of raw records for each station downloaded from ANA web API
(output from \code{\link[=stationsData]{stationsData()}} function).}
}
\value{
A list containing an organized data frame \code{\link[tibble:tibble]{tibble::tibble()}} object
for each station. The data frames will contain the following columns:
station_code: station unique identifier (chr).
consistency_level: data consistency level (1 = raw, 2 = consisted) (chr).
date: date format YYYY-MM-DD (Date).
rainfall_mm/streamflow_m3_s: rain/streamflow gauge measurement (dbl).
}
\description{
Takes as input a list containing data frames of raw records
for each station (output from \code{\link[=stationsData]{stationsData()}}) and organizes
them into tidy data frames. As some dates may be provided with different
consistency levels, priority is given to consisted data (consistency level = 2)
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Fech a inventory of fluviometric stations for the state of Minas Gerais

inv <- inventory(
  states = "MINAS GERAIS",
  stationType = "flu",
  as_sf = TRUE,
  aoi = NULL
)

# Download the first 10 stations from the inventory

s_data <- stationsData(
  inventoryResult = inv[1:10,],
  deleteNAstations = TRUE
)

# Organize the data for the stations

org_data <- organize(
  stationsDataResult = s_data
)
\dontshow{\}) # examplesIf}
}
