% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PettittTest.R
\encoding{UTF-8}
\name{PettittTest}
\alias{PettittTest}
\title{Pettitt trend test}
\usage{
PettittTest(
  dfSeriesFromFillorSerieStatisticsFunc,
  byMonth = FALSE,
  plotGraph = FALSE,
  dirSub = "./pettittTestGraph"
)
}
\arguments{
\item{dfSeriesFromFillorSerieStatisticsFunc}{tibble containing annual or monthly series of all stations;}

\item{byMonth}{logical. if byMounth = TRUE, Pettitt test is performed for each month.}

\item{plotGraph}{logical. defalt = FALSE}

\item{dirSub}{string. directory path to save plots. default = "./pettittTestGraph".}
}
\value{
p-value for each continuous stations data
}
\description{
Performs the Pettitt trend test for annual or monthly rainfall (or streamflow) times series
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

# Fech a inventory of fluviometric stations for the state of Minas Gerais

inv <- inventory(
  states = "MINAS GERAIS",
  stationType = "flu",
  as_sf = TRUE,
  aoi = NULL)

# Download the first 10 stations from the inventory

s_data <- stationsData(
  inventoryResult = inv[1:10,],
  deleteNAstations = TRUE)

# Organize the data for the stations

org_data <- organize(
  stationsDataResult = s_data
)

# Filter the data for desired period and quality contorl

final_data <- selectStations(
  stationsDataResult = org_data,
  mode = "yearly",
  maxMissing = 10,
  minYears = 15,
  month = 1,
  iniYear = NULL,
  finYear = NULL,
  consistedOnly = FALSE,
  plot = TRUE
)

# Annual mean stream flow serie for each station
Qmean_years = seriesStatistics(final_data, statistics = "Qmean")

#MannKendall test
MKTest(dfSeriesFromFillorSerieStatisticsFunc = Qmean_years$df_series, byMonth = FALSE)

\dontshow{\}) # examplesIf}
}
\references{
trend package (https://cran.r-project.org/web/packages/trend/trend.pdf)
}
