% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainStatistics.R
\encoding{UTF-8}
\name{rainStatistics}
\alias{rainStatistics}
\title{Compute rainfall statistics by month or year}
\usage{
rainStatistics(selectStationsResult, statistics = "Rtotal")
}
\arguments{
\item{selectStationsResult}{list, tibble data frame; provides a list containing
the data frames of filtered records for each station
(output from \code{\link[=selectStations]{selectStations()}} function).}

\item{statistic}{character; indicates statistics. The supported statistics are:
(1) total rainfall (Rtotal); (2)  maximum rainfall (Rmax);
(3) rainy days (Rdays). The default is "Rtotal".}
}
\value{
A list containing 3 objects:
\itemize{
\item a list containing statistic a data frame \code{\link[tibble:tibble]{tibble::tibble()}} object for each station.
\item a data frame \code{\link[tibble:tibble]{tibble::tibble()}} with statistic of all stations in wide format
\item a data frame \code{\link[tibble:tibble]{tibble::tibble()}} with statistic of all stations in longer format
}
}
\description{
Takes as input a list containing data frames of organized and filtered records
for each station (output from \code{\link[=selectStations]{selectStations()}}) and compute rainfall statistics
by month or year.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Fecht a inventory of fluviometric stations for the state of Minas Gerais

inv <- inventory(
  states = "MINAS GERAIS",
  stationType = "plu",
  as_sf = TRUE,
  aoi = NULL
)

# Download the first 10 stations from the inventory

s_data <- stationsData(
  inventoryResult = inv[1:10,],
  deleteNAstations = TRUE
)

# Organize the data for the stations

org_data <- organize(
  stationsDataResult = s_data
)

# Filter the data for desired period and quality contorl

final_data <- selectStations(
  stationsDataResult = org_data,
  mode = "yearly",
  maxMissing = 10,
  minYears = 15,
  month = 1,
  iniYear = NULL,
  finYear = NULL,
  consistedOnly = FALSE,
  plot = TRUE
)

# annual mean stream flow serie for each station
Qmean_years = rainStatistics(final_data, statistics = "Rtotal")
\dontshow{\}) # examplesIf}
}
