% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sazonality.R
\name{sazonality}
\alias{sazonality}
\title{Sazonality}
\usage{
sazonality(selectionResultSeries, statistic = "median")
}
\arguments{
\item{selectionResultSeries}{list of tibble; lista de tibbles obtidos com a função \link{selectStations} contendo a série de dados diários.}

\item{statistic}{character; "mean" or "median" monthly statistics.}
}
\value{
dataframe contendo o início do mês úmido e seco.
}
\description{
Identifica os 6 meses secos e úmidos nas estações em análise.
}
\details{
Computa estatística mensal (média ou mediana) nos meses do ano e, posteriormente, e realiza uma média movel de 6 meses para identificar período seco e úmido no ano.
}
\examples{

\dontrun{

# Fech a inventory of fluviometric stations for the state of Minas Gerais

inv <- inventory(
  states = "MINAS GERAIS",
  stationType = "flu",
  as_sf = TRUE,
  aoi = NULL
)

# Download stations from the inventory

s_data <- stationsData(
  inventoryResult = inv[1:10,],
  deleteNAstations = TRUE
)

# Organize the data for the stations

org_data <- organize(stationsDataResult = s_data)

# Filter the data for desired period and quality contorl

final_data <- selectStations(
  organizeResult = org_data,
  mode = "yearly",
  maxMissing = 10,
  minYears = 15,
  month = 1,
  iniYear = NULL,
  finYear = NULL,
  consistedOnly = FALSE,
  plot = TRUE
)

sazon = sazonality(final_data$series, statistic = "median")


}

}
