% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectStationsAllMonths.R
\name{selectStationsAllMonths}
\alias{selectStationsAllMonths}
\title{Select stations loop for all months}
\usage{
selectStationsAllMonths(
  organizeResult,
  mode = "yearly",
  maxMissing = 10,
  minYears = 15,
  iniYear = NULL,
  finYear = NULL,
  consistedOnly = FALSE,
  folderPathWithNameDescription,
  plot = TRUE
)
}
\arguments{
\item{organizeResult}{list, tibble data frame; provides a list containing
the data frames of raw records for each station downloaded from ANA web API
(output from \code{\link[=stationsData]{stationsData()}} function).}

\item{mode}{character; indicates in which scale to check missing data, 'monthly'
or 'yearly'. The default is 'yearly'.}

\item{maxMissing}{numeric; indicates the maximum threshold of missing data allowed.
The default is 10 percent.}

\item{minYears}{numeric; indicates the minimum years of complete data allowed. The
default is 15 years.}

\item{iniYear}{numeric; filters the time series to begin on this year (inclusive).
If you choose to use water year instead of civil year, e.g., month = 6,
the first observation used is from the date "01-06-\code{iniYear}".
The default is NULL (use entire period).}

\item{finYear}{numeric; filters the time series to end on this year (inclusive).
If you choose to use water year instead of civil year, e.g., month = 6,
the last observation used is from the date "31-05-\code{finYear}".
The default is NULL (use entire period).}

\item{consistedOnly}{logical; should only consisted data be considered?
The default is FALSE.}

\item{folderPathWithNameDescription}{character. folder path and description of selectStations parameters.}

\item{plot}{logical; plot the figure? The default is TRUE. The figure is saved
regardless.}
}
\value{
12 Rdata files. Each one contain selectStation result considering inicial hydrological month.
}
\description{
Execute selectStation function for all months (1:12)
}
\examples{

\dontrun{
# Fech a inventory of fluviometric stations for the state of Minas Gerais

inv <- inventory(
  states = "MINAS GERAIS",
  stationType = "flu",
  as_sf = TRUE,
  aoi = NULL
)

# Download the first 10 stations from the inventory

s_data <- stationsData(
  inventoryResult = inv[1:10,],
  deleteNAstations = TRUE
)

# Organize the data for the stations

org_data <- organize(stationsDataResult = s_data)

# Filter the data for desired period and quality contorl

selectStationsAllMonths(
  organizeResult = org_data,
  mode = "yearly",
  maxMissing = 10,
  minYears = 15,
  iniYear = NULL,
  finYear = NULL,
  consistedOnly = FALSE,
  folderPathWithNameDescription = "./loop/selecStation_15years_5porc",
  plot = TRUE
)

#'
files = list.files("./loop", full.names = TRUE) \%>\% gtools::mixedsort()

selectStationsMonth1 = readRDS(files[1])

}
}
