% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationsData.R
\encoding{UTF-8}
\name{stationsData}
\alias{stationsData}
\title{Retrieves raw data for the stations inventory from ANA web API}
\usage{
stationsData(inventoryResult, deleteNAstations = TRUE, waterLevel = FALSE)
}
\arguments{
\item{inventoryResult}{tibble data frame; provides the station inventory (output
from \code{\link[=inventory]{inventory()}} function) for which to download data for.}

\item{deleteNAstations}{logical; should stations with no data be removed?
The default is TRUE.}

\item{waterLevel}{logical. if param "station_type" in \code{\link[=inventory]{inventory()}}, get waterLevel data of fluviometric stations?
Use FALSE for streamflow data. TRUE for water level data. Default is FALSE.}
}
\value{
A list containing a data frame \code{\link[tibble:tibble]{tibble::tibble()}} object for each station.
The data frame format is identical to the format provided by ANA.
}
\description{
Takes as input an inventory of stations (output from
\code{\link[=inventory]{inventory()}}) or data.frame/tibble object with station_code and stationType
columns and downloads raw stations data from the Brazilian
National Water Agency (ANA). The user can choose wether to maintain
stations with missing data.
}
\details{
Improvement of the code developed by Artur Lourenço
(https://github.com/ArturLourenco/HidroWebFix)
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Fech a inventory of fluviometric stations for the state of Minas Gerais

\dontrun{
inv <- inventory(
  states = "MINAS GERAIS",
  stationType = "flu",
  as_sf = TRUE,
  aoi = NULL
)

# Download the first 10 stations from the inventory

s_data <- stationsData(
  inventoryResult = inv[1:10,],
  deleteNAstations = TRUE
)

#######

#create data.frame with station_code and stationType columns

stations_code = data.frame(station_code = c("42600000","42690001"),
                           stationType = "fluviometric")

s_data = stationsData(stations_code)

}



\dontshow{\}) # examplesIf}
}
\references{
Dados Abertos da Agência Nacional de Águas e Saneamento Básico.

\url{https://dadosabertos.ana.gov.br/}

HIDRO - Inventário pluviométrico/fluviométrico atualizado.

\url{https://dadosabertos.ana.gov.br/documents/fb3426be2d4a4f9abfa90fb87b30bd4f/about}
}
