% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terraClimateExtOrg.R
\name{terraClimateExtOrg}
\alias{terraClimateExtOrg}
\title{Extract and organize terraClimate data}
\usage{
terraClimateExtOrg(terraClimateRast, aoi, fun, colname)
}
\arguments{
\item{terraClimateRast}{rast object contain result from downloadterra}

\item{aoi}{object of class \code{sf} or \code{terra} (polygon); Provides the boundaries
where extraction should be done.}

\item{fun}{character; function which should be use on extraction. Available functions are:
min, max, sum, mean, median, mode, majority, minority.}

\item{colname}{chracter; colname of aoi which should be used for display results.}
}
\value{
A list containing an organized data frame \code{\link[tibble:tibble]{tibble::tibble()}} object
for each station. The data frames will contain the following columns:
station_code: station unique identifier (chr).
consistency_level: data consistency level (1 = raw, 2 = consisted) (chr).
date: date format YYYY-MM-DD (Date).
rainfall_mm/streamflow_m3_s: rain/streamflow gauge measurement (dbl).
}
\description{
Extract TerraClimate spatial data based on a polygon and organize it into a time series which can be used at \code{\link[=selectStations]{selectStations()}}
}
\examples{

\dontrun{

es = geobr::read_municipality(code_muni = "ES")

terraClimate = downloadTerraClimate("./temp3",
                                   variable = "ppt",
                                   years = c(2019:2021),
                                   aoi = es)

ppt_es = terraClimateExtOrg(terraClimateRast = terraClimate,
                           aoi = es,
                           fun = "mean",
                           colname = "name_muni")


}
}
