% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watersheDelimit.R
\encoding{UTF-8}
\name{waterShedDelin}
\alias{waterShedDelin}
\title{Delimitation of watersheds based on ANA fluviométric stations}
\usage{
waterShedDelin(
  stationsPath,
  flowAcumPath,
  flowDir8Path,
  bufferSearch = 1000,
  outputDirPath,
  tempDirPath = "./temp"
)
}
\arguments{
\item{stationsPath}{character vector. Shapefile path of ANA fluviometric stations}

\item{flowAcumPath}{character vector. Flow Accumulation raster path}

\item{flowDir8Path}{character vector. Flow direction raster path}

\item{bufferSearch}{value. Search radius in meters to snap pour points}

\item{outputDirPath}{character vector. Output directory path}

\item{tempDirPath}{character vector. Temporary file directory path}
}
\value{
Watersheds delineated
}
\description{
Delimitation of watersheds based on ANA fluviométric stations
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

#

watershedDelimit(stationsPath = "ana_station.shp",
                 flowAcumPath = "flowAccum.tif",
                 flowDir8Path = "./flowDir8.tif",
                 bufferSearch = 1000,
                 outputDirPath = "./watersheds",
                 tempDirPath = "./temp")

\dontshow{\}) # examplesIf}
}
\references{
whitetoolbox package (https://cran.r-project.org/web/packages/whitebox/index.html)
}
