% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershedDelineation.R
\encoding{UTF-8}
\name{watershedDelineation}
\alias{watershedDelineation}
\title{Watershed delineation based on pour points}
\usage{
watershedDelineation(
  stationsPath,
  flowAcumPath,
  flowDir8Path,
  bufferSearch = 1000,
  outputDirPath
)
}
\arguments{
\item{stationsPath}{character vector. Shapefile path of ANA fluviometric stations}

\item{flowAcumPath}{character vector. Flow Accumulation raster path}

\item{flowDir8Path}{character vector. Flow direction raster path}

\item{bufferSearch}{value. Search radius in meters to snap pour points}

\item{outputDirPath}{character vector. Output directory path}
}
\value{
\itemize{
\item nested and unested watersheds in raster and shape format
\item snapped pour points
}
}
\description{
This function delineates watershed boundaries. See details.
}
\details{
The stations shapefile MUST INCLUDE a 'sttn_cd' column containing numerical identifiers for stations.

Shapefiles and rasters MUST BE in the same projection coordinate system (metric).

The delineation could be based on two algorithms:

(1) Pour points are snapped to the chosen flow accumulation pixel value which most approximates the area of the respective station at 'bufferSearch';

(2) Pour points are snapped to the chosen biggest flow accumulation value within a 'bufferSearch' radius.

To choose which algorithm is implemented, the user must configure the input shapefile:

(1) If the first method is desired, the station shapefile MUST INCLUDE an 'are_km2' column containing area values for station basins in km².
If the area is not known for a specific station, leave the 'are_km2' column of this specific station as 'NA', and the function will use the second algorithm to delineate watershed boundaries.

(2) If you're interested in the second algorithm, just make sure that 'are_km2' is not provided ou set 'NA' to all station value at 'are_km2' column.
}
\examples{

\dontrun{
#stations in area of interest

stations = inventory(
  stationType = "flu",
  as_sf = T,
  aoi = sf::st_read("./example/data/aoi_example.shp")
)

# omit station with area equal NA, reproject to epsg of dem and export it

sf::st_write(
  na.omit(stations) \%>\% sf::st_transform(crs = "epsg:32723"),
  dsn = "./example/results/stations_aoi.shp",
  delete_dsn = TRUE, delete_layer = TRUE
)

#run watersheDelimit
watershedDelineation(
  stationsPath = "./example/results/stations_aoi.shp",
  flowAcumPath = "./example/results/demproducts/03flowAccumulation.tif",
  flowDir8Path = "./example/results/demproducts/03flowDirection.tif",
  bufferSearch = 1000,
  outputDirPath = "./example/results/watershedsDelimit"
)

}

}
\references{
whitetoolbox package (https://cran.r-project.org/web/packages/whitebox/index.html)
}
