% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watersheDelimit.R
\encoding{UTF-8}
\name{watersheDelimit}
\alias{watersheDelimit}
\title{(UNDER DEVELOPMENT) Delimitation of watersheds based on ANA fluviométric stations}
\usage{
watersheDelimit(
  stationsPath,
  flowAcumPath,
  flowDir8Path,
  bufferSearch = 1000,
  outputDirPath,
  tempDirPath = "./temp"
)
}
\arguments{
\item{stationsPath}{character vector. Shapefile path of ANA fluviometric stations}

\item{flowAcumPath}{character vector. Flow Accumulation raster path}

\item{flowDir8Path}{character vector. Flow direction raster path}

\item{bufferSearch}{value. Search radius in meters to snap pour points}

\item{outputDirPath}{character vector. Output directory path}

\item{tempDirPath}{character vector. Temporary file directory path}
}
\value{
Watersheds delineated
}
\description{
Delimitation of watersheds based on ANA fluviométric stations
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

#

#stations in area of interest

stations = inventory(
  stationType = "flu",
  as_sf = T,
  aoi = sf::st_read("./example/data/aoi_example.shp")
)

# omit station with area equal NA, reproject to epsg of dem and export it

sf::st_write(
  na.omit(stations) \%>\% sf::st_transform(crs = "epsg:32723"),
  dsn = "./example/results/stations_aoi.shp",
  delete_dsn = TRUE, delete_layer = TRUE
)

#run watersheDelimit
watersheDelimit(
  stationsPath = "./example/results/stations_aoi.shp",
  flowAcumPath = "./example/results/demproducts/03flowAccumulation.tif",
  flowDir8Path = "./example/results/demproducts/03flowDirection.tif",
  bufferSearch = 1000,
  outputDirPath = "./example/results/watershedsDelimit",
  tempDirPath = "./temp"
)

\dontshow{\}) # examplesIf}
}
\references{
whitetoolbox package (https://cran.r-project.org/web/packages/whitebox/index.html)
}
