% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wDelimitationAreaBased.R
\encoding{UTF-8}
\name{wDelimitationAreaBased}
\alias{wDelimitationAreaBased}
\title{Watershed delimitation based on area of ANA stations}
\usage{
wDelimitationAreaBased(
  stationsPath,
  flowAcumPath,
  flowDir8Path,
  bufferSearch = 1000,
  outputDirPath
)
}
\arguments{
\item{stationsPath}{character vector. Shapefile path of ANA fluviometric stations}

\item{flowAcumPath}{character vector. Flow Accumulation raster path}

\item{flowDir8Path}{character vector. Flow direction raster path}

\item{bufferSearch}{value. Search radius in meters to snap pour points}

\item{outputDirPath}{character vector. Output directory path}
}
\value{
\itemize{
\item nested and unested watersheds in raster and shape format
\item snaped pour points
}

#' @references
whitetoolbox package (https://cran.r-project.org/web/packages/whitebox/index.html)
}
\description{
This function delineates watershed boundaries based on the area value displayed in ANA fluviometric station data.
}
\details{
shapefiles and rasters must be at same projection coordinate system (metric)
}
\examples{

\dontrun{
#stations in area of interest

stations = inventory(
  stationType = "flu",
  as_sf = T,
  aoi = sf::st_read("./example/data/aoi_example.shp")
)

# omit station with area equal NA, reproject to epsg of dem and export it

sf::st_write(
  na.omit(stations) \%>\% sf::st_transform(crs = "epsg:32723"),
  dsn = "./example/results/stations_aoi.shp",
  delete_dsn = TRUE, delete_layer = TRUE
)

#run watersheDelimit
watersheDelimit(
  stationsPath = "./example/results/stations_aoi.shp",
  flowAcumPath = "./example/results/demproducts/03flowAccumulation.tif",
  flowDir8Path = "./example/results/demproducts/03flowDirection.tif",
  bufferSearch = 1000,
  outputDirPath = "./example/results/watershedsDelimit"
)

}

}
