% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag1AutoCorrelationAll.R
\name{autocorrelationAll}
\alias{autocorrelationAll}
\title{Lag 1 AutoCorrelation considering all hydrological years months begin}
\usage{
autocorrelationAll(
  selectStationsAllmonthsRDSfolder,
  statistics = "Qmean",
  permanence = 95
)
}
\arguments{
\item{selectStationsAllmonthsRDSfolder}{character. folder path with 12 Rdata files generated with \link{selectStationsAllMonths}.}

\item{statistics}{character; indicates statistics.
\itemize{
\item The supported statistics for streamflow are:
(1) mean stream flow (Qmean); (2)  minimum of seven-day moving average of daily stream flow (Q7);
(3) stream flow associated with a percentage of time (Qperm); (4) maximum stream flow (Qmax);
and (5) minimum stream flow (Qmin).
\item The supported statistics are: (1) total rainfall (Rtotal); (2)  maximum rainfall (Rmax);
(3) rainy days (Rdays).
\item The default value is "Qmean".
}}

\item{permanence}{numeric; percentage of time if "Qperm" is choose as statistic parameter.
The default is 95 percent.}
}
\value{
list with 12 tibble objects. Each tibble contatin autocorrelation for all stations for respective hydrological month begin.
}
\description{
Same processing done at \link{lag1AutoCorrelation} but considering all 12 months possible to start hydrological year
Function idealized to assistant hydrological year definition.
}
\examples{


\dontrun{

# Fech a inventory of fluviometric stations for the state of Minas Gerais.

inv <- inventory(
  states = "MINAS GERAIS",
  stationType = "flu",
  as_sf = TRUE,
  aoi = NULL
)

# Download the first 10 stations from the inventory

s_data <- stationsData(
  inventoryResult = inv[1:10,],
  deleteNAstations = TRUE
)

# Organize the data for the stations

org_data <- organize(
  stationsDataResult = s_data
)

# Filter the data for desired period and quality contorl

selectStationsAllMonths(
  organizeResult = org_data,
  mode = "yearly",
  maxMissing = 10,
  minYears = 15,
  iniYear = NULL,
  finYear = NULL,
  consistedOnly = FALSE,
  folderPathWithNameDescription = "./loop/selecStation_15years_5porc",
  plot = TRUE
)

#autocorrelation for Qmean series for all hydrological year month begin

autoCorQmld = autocorrelationAll(selectStationsAllmonthsRDSfolder = "./loop", statistics = "Qmean")

}
}
