% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demProducts.R
\encoding{UTF-8}
\name{demProducts}
\alias{demProducts}
\title{Obtain digital elevation model products}
\usage{
demProducts(demP, streamP, outputDir)
}
\arguments{
\item{demP}{character vector. Input DEM raster file path.}

\item{streamP}{character vector. Input stream shapefile path.}

\item{outputDir}{character vector. Output rasters file path.}
}
\value{
Create eigth DEM subproducts:
DEM aspect
DEM burned by streams
DEM Sinks
DEM slope in degrees
DEM slop in porcent
DEM burned by streans filled
Flow Accumulation based on filled DEM
Flow Direction based on filled DEM
}
\description{
Obtain digital elevation model products
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

#

demProducts("./dem.tif", "./streams.shp", "./demProducts")

\dontshow{\}) # examplesIf}
}
\references{
whitetoolbox package (https://cran.r-project.org/web/packages/whitebox/index.html)
}
