% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demProducts.R
\encoding{UTF-8}
\name{demProducts}
\alias{demProducts}
\title{(UNDER DEVELOPMENT) Digital elevation model sub-products}
\usage{
demProducts(demP, streamP, outputDir)
}
\arguments{
\item{demP}{character vector. Input DEM raster file path.}

\item{streamP}{character vector. Input stream shapefile path.}

\item{outputDir}{character vector. Output rasters file path.}
}
\value{
Create eigth DEM subproducts:
(1) DEM aspect;
(2) DEM burned by streams;
(3) DEM Sinks;
(4) DEM slope in degrees;
(5) DEM slop in porcent;
(6) DEM burned by streans filled;
(7) Flow Accumulation based on filled DEM;
(8) Flow Direction based on filled DEM.
}
\description{
Obtain digital elevation model sub-products
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

#

demProducts("./dem.tif", "./streams.shp", "./demProducts")

\dontshow{\}) # examplesIf}
}
\references{
whitetoolbox package (https://cran.r-project.org/web/packages/whitebox/index.html)
}
