% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillGaps.R
\encoding{UTF-8}
\name{fillGaps}
\alias{fillGaps}
\title{Fill gaps at monthly or annual time series}
\usage{
fillGaps(StatisticsResult, minimumCor = 0.84, minimunObsPairs = 10)
}
\arguments{
\item{StatisticsResult}{list, tibble data frame; A list containing statistic data frame \code{\link[tibble:tibble]{tibble::tibble()}} object
for each station (output from \code{\link[=flowStatistics]{flowStatistics()}} or \code{\link[=rainStatistics]{rainStatistics()}}).}

\item{minimumCor}{value; minimum correlation between stations. default = 0.84}

\item{minimunObsPairs}{value; minimum of observation pairwise between stations to be filled with.
If \link{StatisticsResult} is annual time series, minimunObsPairs is equal to number of commom years.
If \link{StatisticsResult} is monthly time series, minimunObsPairs is equal to number of common months.}
}
\value{
A list containing 4 objects:
\itemize{
\item a list containing the data frames \code{\link[tibble:tibble]{tibble::tibble()}} for each station after
gap filled.
\item a failureMatrix indicating if the gap was filled (TRUE) or not (FALSE)
\item the saved plot.
}
}
\description{
Takes as input a list containing annual or monthly time series statistic
for each station (output from \code{\link[=flowStatistics]{flowStatistics()}} or \code{\link[=rainStatistics]{rainStatistics()}})
and try to fill gaps based on linear regression among them.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Fech a inventory of fluviometric stations for the state of Minas Gerais

inv <- inventory(
  states = "MINAS GERAIS",
  stationType = "flu",
  as_sf = TRUE,
  aoi = NULL
)

# Download the first 10 stations from the inventory

s_data <- stationsData(
  inventoryResult = inv[1:10,],
  deleteNAstations = TRUE
)

# Organize the data for the stations

org_data <- organize(
  stationsDataResult = s_data
)

# Filter the data for desired period and quality contorl

final_data <- selectStations(
  stationsDataResult = org_data,
  mode = "yearly",
  maxMissing = 10,
  minYears = 15,
  month = 1,
  iniYear = NULL,
  finYear = NULL,
  consistedOnly = FALSE,
  plot = TRUE
)

# Annual mean stream flow serie for each station
Qmean_years = flowStatistics(final_data, statistics = "Qmean")

#fill Gaps of Annual time series

Qmean_years_filled = fillGaps(StatisticsResult = Qmean_years, minimumCor = 0.84, minimunObsPairs = 10)
\dontshow{\}) # examplesIf}
}
