% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyQ7dates.R
\encoding{UTF-8}
\name{identifyQ7dates}
\alias{identifyQ7dates}
\title{Retorna a série de Q7 anual e respectivas datas}
\usage{
identifyQ7dates(selectStationsResultSeries, order = "lastDate")
}
\arguments{
\item{selectStationsResultSeries}{list of tibble data frame;}

\item{order}{character; quando há valores mínimos de Q7 num ano, escolher o primeiro (firstDate) ou último (lastDate).}
}
\value{
A list containing 2 objects:
\itemize{
\item a list of data frames \code{\link[tibble:tibble]{tibble::tibble()}} for each station containing Q7 dates, value and month of ocorrence for each wateryear
\item a dataframe containing Q7 month frequency for each station
}
}
\description{
A partir da série histórica de vazões nas estações (output from \code{\link[=selectStations]{selectStations()}}), retorna os valores de Q7 anual com respectivas datasTakes as input a list containing data frames of organized records
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Fech a inventory of fluviometric stations for the state of Minas Gerais

inv <- inventory(
  states = "MINAS GERAIS",
  stationType = "flu",
  as_sf = TRUE,
  aoi = NULL
)

# Download the first 10 stations from the inventory

s_data <- stationsData(
  inventoryResult = inv[1:10,],
  deleteNAstations = TRUE
)

# Organize the data for the stations

org_data <- organize(stationsDataResult = s_data)

# Filter the data for desired period and quality contorl

final_data <- selectStations(
  organizeResult = org_data,
  mode = "yearly",
  maxMissing = 10,
  minYears = 15,
  month = 1,
  iniYear = NULL,
  finYear = NULL,
  consistedOnly = FALSE,
  plot = TRUE
)


Q7stats = identifyQ7dates( final_data$series)
\dontshow{\}) # examplesIf}
}
