% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKTest.R
\encoding{UTF-8}
\name{MKTest}
\alias{MKTest}
\title{Performs the MK trend test for list of continuous stations data.}
\usage{
MKTest(listStationData, by_month = FALSE)
}
\arguments{
\item{by_month}{logical. if by_mounth = TRUE, MannKendall test is performed for each month.}

\item{dfStationlist}{list with elements containing continuous stations data;}
}
\value{
p-value for each continuous stations data
}
\description{
Performs the MannKendall trend test for list of continuous stations data.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

MKTest(dfStationlist = data, by_month = FALSE)

\dontshow{\}) # examplesIf}
}
\references{
randtest package (https://cran.r-project.org/web/packages/randtests/index.html)
}
