% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKTest.R
\encoding{UTF-8}
\name{MKTest}
\alias{MKTest}
\title{Performs the MannKendall trend test for list of continuous stations data.}
\usage{
MKTest(resultFillorStatistics, by_month = TRUE)
}
\arguments{
\item{resultFillorStatistics}{list with elements containing continuous stations data;}

\item{by_month}{logical. if by_mounth = TRUE, MannKendall test is performed for each month.}
}
\value{
p-value for each continuous stations data
}
\description{
Performs the MannKendall trend test for list of continuous stations data.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

inv <- inventory(
  states = "MINAS GERAIS",
  stationType = "flu",
  as_sf = TRUE,
  aoi = NULL)

# Download the first 10 stations from the inventory

s_data <- stationsData(
  inventoryResult = inv[1:10,],
  deleteNAstations = TRUE)

# Organize the data for the stations

org_data <- organize(
  stationsDataResult = s_data)

# Filter the data for desired period and quality contorl

final_data <- selectStations(
  stationsDataResult = org_data,
  mode = "yearly",
  maxMissing = 10,
  minYears = 15,
  month = 1,
  iniYear = NULL,
  finYear = NULL,
  consistedOnly = FALSE,
  plot = TRUE)

# Annual mean stream flow serie for each station
Qmean_years = flowStatistics(final_data, statistics = "Qmean")

#MannKendall test
MKTest(resultFillorStatistics = Qmean_years, by_month = FALSE)

\dontshow{\}) # examplesIf}
}
\references{
randtest package (https://cran.r-project.org/web/packages/randtests/index.html)
}
