% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectRaster.R
\encoding{UTF-8}
\name{projectRaster}
\alias{projectRaster}
\title{Project raster}
\usage{
projectRaster(raster, projectionEPSG, resolution, rasterP_path)
}
\arguments{
\item{raster}{character vector. Input raster file path.}

\item{projectionEPSG}{character vector. Input "EPSG Geodetic Parameter Dataset"}

\item{resolution}{value. Projected raster resolution in meter.}

\item{rasterP_path}{character vector. Projected output raster file path.}
}
\value{
Projected raster
}
\description{
Project raster to desired EPSG
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

#

projectRaster(raster = "./dem.tif",
              projectionEPSG = "epsg:32723",
              resolution = 30,
              rasterP_path = "./demP.tif")

\dontshow{\}) # examplesIf}
}
\references{
terra package (https://cran.r-project.org/web/packages/terra/index.html)
}
