% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q7Tempiric.R
\encoding{UTF-8}
\name{Q7Tempiric}
\alias{Q7Tempiric}
\title{Compute Q7 associated to desired return period based on empirical distribution
(function design for package internal use. See \link{historicalStatistics} for
general use)}
\usage{
Q7Tempiric(vectorSerieQ7, pReturn = 10)
}
\arguments{
\item{vectorSerieQ7}{numeric vector.}

\item{pReturn}{numeric, desired period of return.}
}
\value{
numeric vector.
}
\description{
Takes as input a vector of Q7 series and Compute Q7 associated to
desired return period.
}
\details{
frequency evaluated with Kimbal method.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Fech a inventory of fluviometric stations for the state of Minas Gerais

inv <- inventory(
  states = "MINAS GERAIS",
  stationType = "flu",
  as_sf = TRUE,
  aoi = NULL
)

# Download the first 10 stations from the inventory

s_data <- stationsData(
  inventoryResult = inv[1:10,],
  deleteNAstations = TRUE
)

# Organize the data for the stations

org_data <- organize(
  stationsDataResult = s_data
)

# Filter the data for desired period and quality contorl

final_data <- selectStations(
  stationsDataResult = org_data,
  mode = "yearly",
  maxMissing = 10,
  minYears = 15,
  month = 1,
  iniYear = NULL,
  finYear = NULL,
  consistedOnly = FALSE,
  plot = TRUE
)

# annual mean stream flow serie for each station
Q7_years = seriesStatistics(final_data, statistics = "Q7")

# Q7 associated with return period of 10 years for first station
Q7.10 = vectorSerieQ7(final_data$series[[1]] \%>\% pull(3), statistics = "Q7")
\dontshow{\}) # examplesIf}
}
