% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunTest.R
\encoding{UTF-8}
\name{RunTest}
\alias{RunTest}
\title{Performs the Wald-Wolfowitz runs test of randomness for list of continuous stations data.}
\usage{
RunTest(dfStationlist, by_month = FALSE)
}
\arguments{
\item{dfStationlist}{list with elements containing continuous stations data;}

\item{by_month}{logical. if by_mounth = TRUE, run test is performed for each month.}
}
\value{
p-value for each continuous stations data
}
\description{
Performs the Wald-Wolfowitz runs test of randomness for list of continuous stations data.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

RunTest(dfStationlist = data, by_month = FALSE)

\dontshow{\}) # examplesIf}
}
\references{
randtest package (https://cran.r-project.org/web/packages/randtests/index.html)
}
