% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sazonality.R
\name{sazonality}
\alias{sazonality}
\title{Sazonality}
\usage{
sazonality(selectionsResultSeries, statistic = "median")
}
\arguments{
\item{selectionsResultSeries}{list of tibble; lista de data frames obtidos com a função selectStations considerando o ano hidrológico igual o civil (month = 1)}

\item{statistic}{character; "mean" or "median" monthly statistics}
}
\value{
dataframe contendo o início do mês úmido e seco.
}
\description{
Identifica os 6 meses secos e úmidos nas estações em análise
}
\details{
Computa estatística mensal (média ou mediana) nos meses do ano e, posteriormente, e realiza uma média movel de 6 meses para identificar período seco e úmido no ano.
}
